+++ Codigo/Componentes/Navegador/EjecuciónNavegación/EjecuciónNavegación/Program.cs
@@ -16,7 +16,7 @@ namespace EjecuciónNavegación
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Vista.Nav());
+            Application.Run(new Vista.navegador());
         }
     }
 }
