--- Codigo/Componentes/Navegador/Vista/Nav.cs
@@ -1,170 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Text;
-using System.Windows.Forms;
-
-namespace Vista
-{
-    public partial class Nav : Form
-    {
-        public Nav()
-        {
-            InitializeComponent();
-            customizeDesing();
-        }
-
-        private void customizeDesing()
-        {
-            //Metodo para cubrir los paneles
-            panelCrud.Visible = false;
-            panelTools.Visible = false;
-        }
-
-        private void hideSubMenu()
-        {
-            if (panelCrud.Visible == true)
-                panelCrud.Visible = false;
-            if (panelTools.Visible == true)
-                panelTools.Visible = false;
-        }
-
-        private void showSubMenu(Panel subMenu)
-        {
-            if (subMenu.Visible == false)
-            {
-                hideSubMenu();
-                subMenu.Visible = true;
-            }
-            else
-                subMenu.Visible = true;
-        }
-
-        private void btnCrud_Click(object sender, EventArgs e)
-        {
-            showSubMenu(panelCrud);
-            
-        }
-
-        private void btnInsert_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnModificar_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnSave_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnCancel_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnDelete_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnSearch_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnUpdate_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnNav_Click(object sender, EventArgs e)
-        {
-            showSubMenu(panelTools);
-           
-        }
-
-        private void btnStart_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnBefore_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnNext_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnEnd_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnReport_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnHelp_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void btnExit_Click(object sender, EventArgs e)
-        {
-            //Codigo
-            //Ocultar submenu
-            hideSubMenu();
-        }
-
-        private void pictureBox2_Click(object sender, EventArgs e)
-        {
-            showSubMenu(panelTools);
-        }
-
-        private void pictureBox14_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void iconPictureBox2_Click(object sender, EventArgs e)
-        {
-
-        }
-    }
-}
