+++ Codigo/Componentes/Navegador/Vista/Vista.csproj
@@ -53,11 +53,11 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Nav.cs">
+    <Compile Include="navegador.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Nav.designer.cs">
-      <DependentUpon>Nav.cs</DependentUpon>
+    <Compile Include="navegador.Designer.cs">
+      <DependentUpon>navegador.cs</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
@@ -94,8 +94,8 @@
     <Content Include="Resources\settings.png" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Nav.resx">
-      <DependentUpon>Nav.cs</DependentUpon>
+    <EmbeddedResource Include="navegador.resx">
+      <DependentUpon>navegador.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
